+++ HISTORY.md
@@ -1,3 +1,11 @@
+
+
+# react-router-dom v6
+* https://velog.io/@soryeongk/ReactRouterDomV6
+* https://reactrouter.com/docs/en/v6/getting-started/overview
+* https://typescript.tv/react/upgrade-to-react-router-v6/
+* https://kkj6670.github.io/board/react/react-router-dom-v6
+
 # 페이지구조
 * mui
 * https://github.com/mui/material-ui/tree/v5.8.3/docs/data/material/getting-started/templates